ml-api/requirements.txt
openai-clip
torchvision
pillow
open_clip_torch
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
aiosqlite==0.21.0
sqlalchemy==2.0.38
greenlet==3.1.1
pydantic-settings==2.7.1
python-multipart==0.0.20
pillow==11.1.0
aiofiles==24.1.0
websockets==15.0
wsproto==1.2.0
uvicorn[standard]==0.34.0